Implemented various permissions

+++ api/company/index.php
@@ -1,13 +1,14 @@
 <?php
-header("Access-Control-Allow-Origin: *");
-header('Content-Type: application/json');
-
+include "../config/authconfig.php";
 include "../database/database.php";
 include "../auth/privileges.php";
 include "./get.php";
 include "../saved/get.php";
 include "./post.php";
 include "./delete.php";
+
+header("Access-Control-Allow-Origin: $cors");
+header('Content-Type: application/json');
 
 if(!isset($_POST["user"])) {
   echo json_encode(array(
@@ -16,18 +17,18 @@ if(!isset($_POST["user"])) {
   ));
   die();
 }
-
-$user = $_POST['user'];
-if(!hasPermission($user, $_POST['REQUEST_METHOD'])) {
-  echo json_encode(array(
-    "error" => true,
-    "message" => "{$_POST['user']} doesn't have permission to {$_POST['REQUEST_METHOD']}"
-  ));
-  die();
-}
+$user = intval($_POST["user"]);
 
 switch ($_POST["REQUEST_METHOD"]) {
   case "GET":
+    if(!hasPermission($user, "BASE")) {
+      echo json_encode(array(
+        "error" => true,
+        "message" => "The given user does not have BASE permissions."
+      ));
+      die();
+    }
+
     if(isset($_POST["id"])) {
       $company = getCompanyById($_POST["id"]);
       $company["saved"] = isSavedBy($user, $company["id"]);
@@ -55,6 +56,14 @@ switch ($_POST["REQUEST_METHOD"]) {
     break;
 
   case "POST":
+    if(!hasPermission($user, "MANAGE_COMPANY")) {
+      echo json_encode(array(
+        "error" => true,
+        "message" => "The given user does not have MANAGE_COMPANY permissions."
+      ));
+      die();
+    }
+
     if(isset($_POST["name"]) && strlen($_POST["name"]) > 0 && isset($_POST["fields"])) {
       echo json_encode(
         insertCompany(
@@ -66,10 +75,25 @@ switch ($_POST["REQUEST_METHOD"]) {
     break;
 
   case "PUT":
-    // code...
+    if(!hasPermission($user, "MANAGE_COMPANY")) {
+      echo json_encode(array(
+        "error" => true,
+        "message" => "The given user does not have MANAGE_COMPANY permissions."
+      ));
+      die();
+    }
+
     break;
 
   case "DELETE":
+    if(!hasPermission($user, "MANAGE_COMPANY")) {
+      echo json_encode(array(
+        "error" => true,
+        "message" => "The given user does not have MANAGE_COMPANY permissions."
+      ));
+      die();
+    }
+
     if(isset($_POST["id"])) {
       deleteCompanyById($_POST["id"]);
     }
